게시물 수정 기능 추가

+++ sangilBlog/src/main/java/com/sang12/blog/repository/common/BoardRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sang12.blog.domain.board.BoardEntity;
 
-public interface BoardRepository  extends JpaRepository<BoardEntity, Integer>{ 
+public interface BoardRepository  extends JpaRepository<BoardEntity, Long>{ 
 	Page<BoardEntity> findAllByOrderByBoardId(Pageable pageable);
 }
